+++ apps/dialtone-documentation/docs/.vuepress/baseComponents/DialtoneUsage.vue
@@ -2,8 +2,10 @@
   <div class="dialtone-usage">
     <div class="dialtone-usage__item dialtone-usage__item--do">
       <h3 class="dialtone-usage__hd dialtone-usage__hd--do">
-        <dt-icon name="thumbs-up" size="400" />
-        Do
+        <dt-icon name="check" size="200" />
+        <div class="dialtone-usage__label">
+          Do
+        </div>
       </h3>
       <div class="dialtone-usage__bd">
         <slot name="do" />
@@ -11,8 +13,10 @@
     </div>
     <div class="dialtone-usage__item dialtone-usage__item--dont">
       <h3 class="dialtone-usage__hd dialtone-usage__hd--dont">
-        <dt-icon name="thumbs-down" size="400" />
-        Don’t
+        <dt-icon name="close" size="200" />
+        <div class="dialtone-usage__label">
+          Don’t
+        </div>
       </h3>
       <div class="dialtone-usage__bd">
         <slot name="dont" />