+++ packages/dialtone-tokens/build-android.js
@@ -4,5 +4,6 @@
  * Builds the android tokens to the correct final package format (output to dist_android)
  */
 
-const mvn = require('maven').create();
-mvn.execute(['package'])
+import Maven from 'maven';
+const mvn = Maven.create();
+mvn.execute(['package']);