+++ packages/dialtone-vue3/.storybook/preview.jsx
@@ -1,10 +1,14 @@
 import '../css/dialtone-globals.less';
 import '@dialpad/dialtone-css/lib/dist/dialtone.css';
+import { addons } from '@storybook/preview-api';
+import { setTheme } from '@/../../common/themes/config';
+import DpLight from '@/../../common/themes/dp-light.js';
+import DpDark from '@/../../common/themes/dp-dark.js';
 import { MINIMAL_VIEWPORTS } from '@storybook/addon-viewport';
 import { setup } from '@storybook/vue3';
 import React from 'react';
 import { DocsContainer } from '@storybook/addon-docs';
-import { useDarkMode } from "storybook-dark-mode";
+import { useDarkMode, DARK_MODE_EVENT_NAME } from "storybook-dark-mode";
 import fixDefaultSlot from '../components/plugins/fixDefaultSlot';
 import { setEmojiAssetUrlSmall, setEmojiAssetUrlLarge, setCustomEmojiUrl, setCustomEmojiJson } from '@/common/emoji';
 import customEmojiJson from '@/common/custom-emoji.json';
@@ -12,6 +16,12 @@ import { dialtoneDarkTheme, dialtoneLightTheme } from './dialtone-themes.js';
 import { DtTooltipDirective } from "@/directives/tooltip";
 import { faker } from '@faker-js/faker';
 import { DtScrollbarDirective } from "@/directives/scrollbar";
+
+const channel = addons.getChannel();
+
+channel.on(DARK_MODE_EVENT_NAME, (isDark) => {
+  setTheme(isDark ? DpDark : DpLight);
+});
 
 setEmojiAssetUrlSmall('https://static.dialpadcdn.com/joypixels/png/unicode/32/', '.png');
 setEmojiAssetUrlLarge('https://static.dialpadcdn.com/joypixels/svg/unicode/', '.svg');
@@ -82,13 +92,6 @@ export default {
       },
     },
     backgrounds: { disable: true },
-    darkMode: {
-      darkClass: 'dialtone-theme-dark',
-      lightClass: 'dialtone-theme-light',
-      dark: dialtoneDarkTheme,
-      light: dialtoneLightTheme,
-      stylePreview: true,
-    },
     docs: {
       container: ({ children, ...props }) => {
         const isDark = useDarkMode();