+++ packages/dialtone-vue3/recipes/cards/ivr_node/ivr_node.stories.js
@@ -176,6 +176,14 @@ export const Branch = {
   },
 };
 
+export const Assign = {
+  render: DefaultTemplate,
+
+  args: {
+    nodeType: 'assign',
+  },
+};
+
 export const Transfer = {
   render: DefaultTemplate,