+++ packages/dialtone-vue3/recipes/comboboxes/combobox_multi_select/combobox_multi_select_default.story.vue
@@ -14,6 +14,7 @@
     :selected-items="$attrs.selectedItems"
     :max-selected="$attrs.maxSelected"
     :list-max-height="$attrs.listMaxHeight"
+    :list-max-width="$attrs.listMaxWidth"
     :collapse-on-focus-out="$attrs.collapseOnFocusOut"
     :max-selected-message="$attrs.maxSelectedMessage"
     :has-suggestion-list="$attrs.hasSuggestionList"
@@ -25,6 +26,7 @@
     @select="onComboboxSelect"
     @remove="onComboboxRemove"
     @max-selected="$attrs.onMaxSelected"
+    @combobox-highlight="$attrs.onComboboxHighlight"
   >
     <template
       v-if="$attrs.header"