+++ packages/dialtone-tokens/CHANGELOG.md
@@ -1,3 +1,10 @@
+# [1.33.0](https://github.com/dialpad/dialtone/compare/dialtone-tokens/v1.32.0...dialtone-tokens/v1.33.0) (2024-08-01)
+
+
+### Features
+
+* **Tokens:** DLT-1696 theme and brand tokens ([#416](https://github.com/dialpad/dialtone/issues/416)) ([11085dc](https://github.com/dialpad/dialtone/commit/11085dc9aeb0698f8f9f277c9b66b9f5869bec71))
+
 # [1.32.0](https://github.com/dialpad/dialtone/compare/dialtone-tokens/v1.31.1...dialtone-tokens/v1.32.0) (2024-07-12)